requirements.txt
asgiref==3.8.1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==4.1.13
django-extensions==3.2.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
firebase-admin==6.6.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
networkx==3.4.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.3
PyJWT==2.10.1
pymongo==3.11.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
rsa==4.9
scikit-fuzzy==0.5.0
scipy==1.11.1
six==1.17.0
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
